+++ database/migrations/2018_03_06_022856_create_comments_table.php
@@ -21,7 +21,7 @@ class CreateCommentsTable extends Migration
             $table->string('commentable_type');
             $table->unsignedInteger('parent_id')->nullable();
             $table->smallInteger('level')->default(1);
-            $table->integer('is_hidden')->default(0);
+            $table->smallInteger('is_hidden')->default(0);
             $table->timestamps();
         });
     }
